Create README.md
To get a free battle pass in fortnite is quite simple. First you are going to have to steal a car. It has to be a strong, fast, hypercar like a mini cooper or something. Then you will have to steal 60 bucks from your parents for gas because this is going to be a long ride. Unless you live in Cary, North Carolina. One more thing you will need to be prepared for this mission is you have to steal an automatic rifle. Make sure the rifle is loaded and you have plenty of ammo. Then you will have to drive to Cary, North Carolina. Once you get there, go get some gas. If you don't need gas, siphon the gas from the car and put it into a gas tank. You will need it later. One more thing you will need is some matches. This will be used with the gas. Then you will go to Epic Games Headquarters. There will be 4 security guards out there protecting the place so you will have to take them down. Next, you will enter the headquarters and you will see lots of cameras and a security check. You don't have to worry about the cameras because you will burn the headquarters down later on. Next, there will be six more security guards and some workers. Take down the security guards and take the workers with you to the elevator. Once you get to the elevator, go inside and go to floor 12. This is where the fun part starts. You will shoot the ceiling to make the workers feel threatened and they should go on the floor. Next you will kill the security guards. Then you will keep walking until you find some white guy with a suit and a vintage rolex watch. Pick up the guy and take him to the elevator. Once you get to the elevator, go all the way to the top floor. When you get to the top floor, tell him to go into my fortnite account and give me the battle pass. If he refuses, hit him in the head with your gun with 75% of your strength. If he still refuses, shoot his foot. This should get him to do it. When he transfers the battle pass into your account you will go into the elevator. Go all the way to the bottom floor. Drop all the gas on the elevator floor and get out of the elevator. Quickly light up a match and throw it on top of the gas and run. Good job, you just got yourself a battle pass. Enjoy.